requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
pyyaml==6.0
pillow==10.0.0
